Fixed delta time

+++ CaveKingdom/source/input/input.c
@@ -138,7 +138,7 @@ Vector2 from_screen_to_tile_coords(Vector2 screen_coords) {
 
     Vector2f adjusted_coords = vector2f_sum(screen_coords_f, camera_f);
 
-    Vector2f tile_size_f = { TILE_SIZE * camera.zoom, TILE_SIZE * camera.zoom };
+    Vector2f tile_size_f = { (float)(TILE_SIZE * camera.zoom), (float)(TILE_SIZE * camera.zoom) };
 
     Vector2f tile_coords_f = vector2f_divide(adjusted_coords, tile_size_f);
